+++ src/main/java/com/boldfaced7/board/dto/request/AuthRequest.java
@@ -1,6 +1,10 @@
 package com.boldfaced7.board.dto.request;
 
+import com.boldfaced7.board.domain.Member;
 import com.boldfaced7.board.dto.AuthDto;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,7 +15,14 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AuthRequest {
+
+    @Email
+    @NotBlank
+    @Size(max = Member.MAX_EMAIL_LENGTH)
     private String email;
+
+    @NotBlank
+    @Size(max = Member.MAX_PASSWORD_LENGTH)
     private String password;
 
     public AuthDto toDto() {
