+++ src/main/java/com/boldfaced7/board/dto/request/SaveArticleRequest.java
@@ -1,7 +1,9 @@
 package com.boldfaced7.board.dto.request;
 
+import com.boldfaced7.board.domain.Article;
 import com.boldfaced7.board.dto.ArticleDto;
-import com.boldfaced7.board.dto.response.AuthResponse;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,25 +13,20 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ArticleRequest {
+public class SaveArticleRequest {
+
+    @NotBlank
+    @Size(max = Article.MAX_TITLE_LENGTH)
     private String title;
+
+    @NotBlank
+    @Size(max = Article.MAX_CONTENT_LENGTH)
     private String content;
-    private String author;
 
     public ArticleDto toDto() {
         return ArticleDto.builder()
                 .title(title)
                 .content(content)
-                .author(author)
-                .build();
-    }
-
-    public ArticleDto toDto(Long targetId) {
-        return ArticleDto.builder()
-                .articleId(targetId)
-                .title(title)
-                .content(content)
-                .author(author)
                 .build();
     }
 }
